moved TopMenu to DefaultLayout
