+++ frontend/src/layouts/DefaultLayout.vue
@@ -1,6 +1,35 @@
 <template>
   <v-main class="bg-background main-layout">
-    <TopMenu />
+    <!-- Top Menu -->
+    <v-app-bar flat class="app-bar" height="70">
+      <v-row class="ma-1">
+        <v-col class="d-none d-md-flex align-center">
+          <a href="/" class="logo">
+            <LogoImage size="small" />
+          </a>
+        </v-col>
+        <v-col class="d-flex align-center justify-center">
+          <TabControl />
+        </v-col>
+        <v-col class="d-none d-md-flex align-center">
+          <v-row>
+            <v-col class="d-flex align-center">
+              <LightDarkSwitch class="d-none d-lg-flex" />
+            </v-col>
+            <v-col class="d-flex align-center justify-end">
+              <button @click="toggleDrawer">
+                <Circle>
+                  <v-icon icon="$camera"></v-icon>
+                </Circle>
+              </button>
+              <UserInfo class="ml-2" />
+            </v-col>
+          </v-row>
+        </v-col>
+      </v-row>
+    </v-app-bar>
+    <TablesDrawer v-model="isTablesDrawerVisible" class="hide-on-mobile" location="right" />
+
     <v-container fluid class="page-container px-8">
       <div class="create-button-container">
         <div>
@@ -57,10 +86,21 @@ import { ref } from 'vue'
 import LargeDreamMallButton from '#components/buttons/LargeDreamMallButton.vue'
 import MainButton from '#components/buttons/MainButton.vue'
 import BottomMenu from '#components/menu/BottomMenu.vue'
-import TopMenu from '#components/menu/TopMenu.vue'
+import Circle from '#components/menu/CircleElement.vue'
+import LightDarkSwitch from '#components/menu/LightDarkSwitch.vue'
+import LogoImage from '#components/menu/LogoImage.vue'
+import TabControl from '#components/menu/TabControl.vue'
+import UserInfo from '#components/menu/UserInfo.vue'
+import TablesDrawer from '#components/tablesDrawer/TablesDrawer.vue'
 import { JoinMyTableMutationResult, joinMyTableMutation } from '#mutations/joinMyTableMutation'
 import GlobalErrorHandler from '#plugins/globalErrorHandler'
 import { useActiveTableStore } from '#stores/activeTableStore'
+
+const isTablesDrawerVisible = defineModel<boolean>()
+
+const toggleDrawer = () => {
+  isTablesDrawerVisible.value = !isTablesDrawerVisible.value
+}
 
 const isButtonListVisible = ref(false)
 
@@ -115,6 +155,20 @@ const enterTable = async () => {
     .page-container {
       margin-bottom: 50px;
     }
+  }
+}
+
+.app-bar {
+  background: transparent !important;
+}
+
+.logo {
+  width: 140px;
+}
+
+.hide-on-mobile {
+  @media #{map-get($display-breakpoints, 'md-and-down')} {
+    display: none;
   }
 }
